src/fibonacci/requirements.txt
PyMySQL==0.8.0
django-cors-headers==2.2.0
djangorestframework==3.6.2
PyMySQL==0.8.0
pytz==2018.3